move adapters, ipfs client, interfaces to utils
