+++ packages/app/components/Proposals/Voting/OpenVoting.tsx
@@ -2,7 +2,7 @@ import { RadioGroup } from '@headlessui/react';
 import { setDualActionModal } from 'context/actions';
 import { store } from 'context/store';
 import { ContractsContext } from 'context/Web3/contracts';
-import { Proposal, ProposalType, Status } from 'interfaces/interfaces';
+import { Proposal, ProposalType, Status } from '@popcorn/utils/';
 import { useContext, useState } from 'react';
 
 enum VoteOptions {
