+++ packages/app/pages/beneficiaries/[id].tsx
@@ -2,9 +2,8 @@ import { ContractsContext } from 'context/Web3/contracts';
 import { useRouter } from 'next/router';
 import { useContext, useEffect, useState } from 'react';
 import BeneficiaryPage from 'components/Beneficiaries/BeneficiaryPage';
-import { BeneficiaryApplication } from '../../interfaces/interfaces';
-import { IpfsClient } from 'utils/IpfsClient';
-import { BeneficiaryAdapter } from 'utils/BeneficiaryAdapter';
+import { BeneficiaryApplication, IpfsClient, BeneficiaryRegistryAdapter } from '@popcorn/utils/';
+
 export default function SingleBeneficiaryPage(): JSX.Element {
   const router = useRouter();
   const { contracts } = useContext(ContractsContext);
@@ -12,7 +11,7 @@ export default function SingleBeneficiaryPage(): JSX.Element {
 
   useEffect(() => {
     if (contracts) {
-      BeneficiaryAdapter(contracts.beneficiary, IpfsClient)
+      BeneficiaryRegistryAdapter(contracts.beneficiary, IpfsClient)
         .getBeneficiaryApplication(router.query.id as string)
         .then((beneficiaryApplication) =>
           setBeneficiary(beneficiaryApplication),
