+++ packages/app/pages/beneficiaries/index.tsx
@@ -1,19 +1,16 @@
 import BeneficiaryGrid from 'components/Beneficiaries/BeneficiaryGrid';
-import { BeneficiaryApplication } from 'interfaces/interfaces';
 import { useContext, useEffect, useState } from 'react';
-import { BeneficiaryAdapter } from 'utils/BeneficiaryAdapter';
-import { IpfsClient } from 'utils/IpfsClient';
 import { ContractsContext } from '../../context/Web3/contracts';
+import { BeneficiaryApplication, IpfsClient, BeneficiaryRegistryAdapter } from '@popcorn/utils/';
 
 export default function BeneficiaryPage(): JSX.Element {
   const { contracts } = useContext(ContractsContext);
   const [beneficiaries, setBeneficiaries] = useState<BeneficiaryApplication[]>(
     [],
   );
-
   useEffect(() => {
     if (contracts) {
-      BeneficiaryAdapter(contracts.beneficiary, IpfsClient)
+      BeneficiaryRegistryAdapter(contracts.beneficiary, IpfsClient)
         .getAllBeneficiaryApplications()
         .then((beneficiaries) => setBeneficiaries(beneficiaries));
     }
