+++ packages/app/pages/beneficiary-proposals/takedowns/[id].tsx
@@ -1,5 +1,5 @@
+import { ProposalType } from '@popcorn/utils/';
 import ProposalPage from 'components/Proposals/ProposalPage';
-import { ProposalType } from 'interfaces/interfaces';
 
 export default function SingleTakedownPage(): JSX.Element {
   return <ProposalPage proposalType={ProposalType.Takedown} />;
