+++ packages/app/pages/beneficiary-proposals/takedowns/index.tsx
@@ -1,5 +1,5 @@
+import { ProposalType } from '@popcorn/utils/';
 import ProposalGrid from 'components/Proposals/ProposalGrid';
-import { ProposalType} from 'interfaces/interfaces';
 
 export default function TakedownPage(): JSX.Element {
   return <ProposalGrid proposalType={ProposalType.Takedown} />;
