+++ packages/utils/src/index.ts
@@ -1,4 +1,10 @@
 export { default as logger } from './logger';
 export { default as log } from './logger';
 export { default as useFetch } from './useFetch';
-export {  GrantElectionAdapter } from "./Contracts";
+export { GrantElectionAdapter } from './Contracts';
+export { Status, ProposalType } from './interfaces/interfaces';
+export type { BeneficiaryApplication, Proposal } from './interfaces/interfaces';
+export { IpfsClient } from './IpfsClient/IpfsClient';
+export { BeneficiaryRegistryAdapter } from './Contracts/BeneficiaryRegistry/BeneficiaryRegistryAdapter';
+export { BeneficiaryGovernanceAdapter } from './Contracts/BeneficiaryGovernance/BeneficiaryGovernanceAdapter';
+
